pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "split-inference"
version = "0.1.0"
description = "Privacy-Preserving Split LLM Inference"
readme = "split_inference/README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "torch>=2.1",
    "transformers>=4.40",
    "grpcio>=1.60",
    "grpcio-tools>=1.60",
    "numpy",
    "protobuf>=4.25",
]

[project.optional-dependencies]
server = [
    "sglang>=0.4",
]
dev = [
    "pytest",
    "pytest-asyncio",
    "grpcio-testing",
]

[project.scripts]
split-local = "split_inference.local_server.client:main"
split-server = "split_inference.main_server.launch:main"
split-gencerts = "split_inference.crypto.channel:generate_certificates"

[tool.setuptools.packages.find]
include = ["split_inference*"]
